# Optimize Zthuee projectiles (amongst other projectile things)