--- a/engine/Sim/Projectile.lua
+++ b/engine/Sim/Projectile.lua
@@ -18,6 +18,7 @@ end
 
 ---
 --  Projectile:CreateChildProjectile(blueprint)
+-- Internally calls import to find the projectile
 function Projectile:CreateChildProjectile(blueprint)
 end