--- a/lua/sim/Profiler.lua
+++ b/lua/sim/Profiler.lua
@@ -19,7 +19,7 @@ local thread = false
 --- Data that we send over to the UI
 local data = CreateEmptyProfilerTable()
 
-local lastImport = ""
+local traces = { }
 
 --- Toggles the profiler on / off
 function ToggleProfiler(army, forceEnable)
@@ -105,12 +105,14 @@ function ToggleProfiler(army, forceEnable)
                     scope = "other"
                 end
 
-                if name == "import" then 
+                if name == "lambda" then 
                     local trace = repr(debug.traceback())
-                    if lastImport ~= trace then 
-                        lastImport = trace 
-                        LOG(trace)
+                    if not traces[trace] or traces[trace] == 500 then
+                        traces[trace] = traces[trace] or 0  
+                        LOG(tostring(traces[trace]) .. ": " .. trace)
                     end
+                    
+                    traces[trace] = traces[trace] + 1
                 end
 
                 -- keep track