--- a/lua/sim/Unit.lua
+++ b/lua/sim/Unit.lua
@@ -1031,14 +1031,16 @@ Unit = Class(moho.unit_methods) {
     -------------------------------------------------------------------------------------------
 
     OnDamage = function(self, instigator, amount, vector, damageType)
-        if self.CanTakeDamage then
-            self:DoOnDamagedCallbacks(instigator)
+        if damageType ~= "KnockTree" then 
+            if self.CanTakeDamage then
+                self:DoOnDamagedCallbacks(instigator)
 
-            -- Pass damage to an active personal shield, as personal shields no longer have collisions
-            if self:GetShieldType() == 'Personal' and self:ShieldIsOn() and not self.MyShield.Charging then
-                self.MyShield:ApplyDamage(instigator, amount, vector, damageType)
-            else
-                self:DoTakeDamage(instigator, amount, vector, damageType)
+                -- Pass damage to an active personal shield, as personal shields no longer have collisions
+                if self:GetShieldType() == 'Personal' and self:ShieldIsOn() and not self.MyShield.Charging then
+                    self.MyShield:ApplyDamage(instigator, amount, vector, damageType)
+                else
+                    self:DoTakeDamage(instigator, amount, vector, damageType)
+                end
             end
         end
     end,