--- a/lua/system/import.lua
+++ b/lua/system/import.lua
@@ -36,20 +36,19 @@ local StringSub = string.sub
 -- these values can be adjusted by hooking into this file
 local informDevOfLoad = false
 
--- local once = true
-
 --- The global import function used to keep track of modules.
 -- @param name The path to the module to load.
 function import(name)
 
-    -- if once then 
-    --     once = false
-    --     LOG(repr(debug.listcode(import)))
-    -- end
-
-    -- caching: if it exists then we return the previous version
-    name = StringLower(name)
+    -- attempt to find the module without lowering the string
     local existing = upModules[name]
+    if existing then
+        return existing
+    end
+
+    -- try again after lowering the string
+    name = StringLower(name)
+    existing = upModules[name]
     if existing then
         return existing
     end