--- a/projectiles/SIFThunthoArtilleryShell01/SIFThunthoArtilleryShell01_script.lua
+++ b/projectiles/SIFThunthoArtilleryShell01/SIFThunthoArtilleryShell01_script.lua
@@ -46,7 +46,7 @@ SIFThunthoArtilleryShell01 = Class(SThunthoArtilleryShell) {
     end,
 
     OnImpact = function(self, TargetType, TargetEntity) 
-    
+        LOG("OnImpact!")
         local Random = Random
 
         -- the split fx