Our benchmark is 400 Zthuee firing at once on the map White Fire. The map has no civilians. All functionality of the original projectile is still there, just faster. A base line is the same scenario, but without the Zthuee firing.

On the FAF branch: `12 - 14 ms` with a base line of `3 ms`
![here-is-gif-15-faf](https://user-images.githubusercontent.com/15778155/160083873-09dda77e-42cb-4867-b72f-5a2fa2572da5.gif)

This pull request: `9 - 12 ms` with a base line of `3 ms`
![here-is-gif-15-develop](https://user-images.githubusercontent.com/15778155/160083889-9fcc912f-79c6-4a69-89eb-9f308a460065.gif)

This pull request, without decals: `9 - 10.5 ms` with a base line of `3 ms`
![here-is-gif-15-develop-no-decals](https://user-images.githubusercontent.com/15778155/160083895-9e989399-b6fd-41ac-9271-ea2a4c0a7810.gif)

If we'd reduce the base lines:
 - FAF branch: `9 - 11 ms` for firing
 - This pull request: `6 - 9 ms` for firing
 - This pull request, without decals: `6 - 7.5 ms` when firing
 
Therefore we can safely assume that there's at least a 20% performance improvement. Some of these changes apply to all projectiles.